Added Trainer Database column
Added the RewardAmount column in the trainer Database with base amounts for defeating each trainer.

Added the ability to store and fetch the RewardAmount from the database in the trainer and trainerBuilder files.

Made edits to the canComplete method in trainer.

Added a plus or minus 10 percent token bonus when calculating the tokens earned after a battle.