Finished adding Getter Tests
Changed the abilities interface so that getAbilityName() method is now the toString() method. This change was made so that test reporting would look nice. Plus we should have just used toString() from the start.

Added a toString() method to NullMoveSet.java so that test reporting looks nice.

Changed NullStats.java to extend Stats instead of Implementing IStats. This was so that we don't have a file full of methods that will ever be used. I also added a toString() method to this file so that testing reporting looks nice.

I added a getClone()  method to Stats.java. This is in preparation for evolutions but it also helped me make testing easier.

I added methods to StatsBuilder,java so that during a stats cloning process we can port over EV points after an anature has leveled up to the correct level. This also fixes the problem of overpowered evolutions. I then renamed a variable from mStatsCore to mStats.

Found a bug in Anature where is a setter method is checking against a different instance of NullAbility.

I removed the NullAbility from AbilityPool.java and added a method so that I can grab every ability for testing purposes.

I made a significant change to the TestObjects.java file for testing reasons. Making ever default variable accessible for speeding up the process of making test files int he future.

Lastly I refactored the Setter Tests majorly and made them very dynamic using annotations form the JUnit api. I organized the tests so that test reports are very easy to read. During this process I also organized the test file so that navigation is simple which should speed up our ability to work with the tests for updates or discovery reasons. I also added Tests for the Getter methods.