Added method for Nature value
Refactor in Stats so that the Nature modifier amount can be accessed via a method. During this processes refactored the code.

Added a Enum class Stat that has all the stats of a Anature.

Made changes to AnatureSummaryController for testing.

Edited Natures to use new Enum and deleted the old nested Enum.

Edited Stats interface to expose new method call.