Refactoring Moves
Move.Java:
Turned this class back to being abstract because we can nest classes now apparently in java so I am using a new way of parsing moves which can be seen in MoveCollection,java. Changed the activate Move back to being abstract as well.

MoveBuilder.java:
Got rid of the hideous switch statement that figures out what version the move is going to be. Removed Generics on MoveBuilder because Move.java is now abstract. Replaced switch statement with a call to get the the move from the MoveCollection.java file.

MoveCollection.java:
Added a new file called MoveColection.java that is the main hub for all move logic. Within hold all the version of Moves that we have via classes. Those classes link to methods that hold all the logic that move needs to execute. Complex Moves that could not be abstracted or were not are at the bottom of the file. The benefit here is that we can see every move all in one file now so when we make a correction it effects all the moves. When we make a new decision on how to change something we can reevaluate this file. This change way made mainly to clean up the MoveBuilder but then it ended up just being a main hub.

Moves*:
Moved all move logic into the  MoveCollection.java file. 

NullMoveSet.java was updated to use the NullMove from  MoveCollection.

StatsTemp.java got an update to support the new Stat enum. Four methods were added to support incrementing or decrementing a temp stat(s). This was used heavily in the new MoveCollection.java file.

Anature.java added the method call to port over the catch rate variable in getClone() method

AnatureBuilder.java made the withCatchRate() Method public so we can use it.

MovePool.java removed the Generic declaration to form to the new requirements.

TestObjects.java edited the TestObjects FIle so I could run the tests.