Fully implemented Saving and Loading
To save the game click on the "save" side menu button.

To load a save click on the "quit" side menu button.

Startup.java is where all the magic happens. Created a interface SavableItem and an enum that implements it called SaveItem. This enum hold all the objects that will be saved int he game Each enum should implement the methods in SavableItem where the item and loading logic will take place. Besides that two methods were added called load() and save() which load and save the game.

AbstractController.java added a method called saveLoadUpdates() which is called on save and on laod.

SideMenuController.java the save button saves the game and the quit button loads the game. Both will set the side menu visibility to false as well.

All the other changes were to add Serializable to the requred files.

Please note that there is only one save file as of right now. If we want to add the ability to ad more save file files that should be done in a different case.